api_endpoints/requirements.txt
asgiref==3.4.1
CD==1.1.0
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0
-e git+https://github.com/django-extensions/django-extensions.git@f67ff680cd6d7264cdce05309b537ac2e1ee4a70#egg=django_extensions
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
drf-yasg==1.20.0
Faker==11.3.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==21.3
pipreqs==0.4.11
psycopg2==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
yarg==0.1.9
